Factor out MediaContainer and Playlist page sections to their own components
